strnote.cppでエラーが出るのを修正 HSP3Dish iOS 64bit仮対応
git-svn-id: http://dev.onionsoft.net/svn/openhsp/trunk@681 ceb3ebad-0e53-d446-8596-a45b2b8d79cd

+++ package/win32/iOS/iHSP20admob/hsp3/hsp3int.cpp
@@ -1096,13 +1096,13 @@ static void *reffunc_intfunc( int *type_res, int arg )
 		STRUCTDAT *st;
 		if ( *type == TYPE_DLLFUNC ) {
 			st = &(ctx->mem_finfo[ *val ]);
-			reffunc_intfunc_ivalue = (int)(st->proc);
+			reffunc_intfunc_ivalue = (int)(size_t)(st->proc);
 			code_next();
 			break;
 		}
 		aptr = code_getva( &pval );
 		pdat = HspVarCorePtrAPTR( pval, aptr );
-		reffunc_intfunc_ivalue = (int)(pdat);
+		reffunc_intfunc_ivalue = (int)(size_t)(pdat);
 		break;
 		}
 	case 0x00d:								// varuse
